Mostly proof of concept work
For the moment, this code includes a robot picture which travels across
the screen and ends the game when it hits the player. Each paint event
ups the score by 1, which makes it loosely attached to game time.
pressing down is also worth a point. this was mostly to test my
highscore code. The actual robot and scoring system will still need
overhauled eventually. Bottom line is that changes to main window and
player are important, but the work in game could probably be ignored.